refactor(chat): use layout to avoid refetching all chats on every chat
