+++ frontend/app/chat/hooks/useChats.ts
@@ -1,0 +1,8 @@
+import { useState } from "react";
+import { Chat } from "../types";
+
+export default function useChats() {
+  const [chats, setChats] = useState<Chat[]>([]);
+
+  return { chats, setChats };
+}
