--- frontend/app/chat/page.tsx
@@ -1,4 +1,0 @@
-"use client";
-import ChatPage from "./[...chatId]/page";
-
-export default ChatPage;
